feat: go mode
